+++ spp_hide_menus_base/tests/__init__.py
@@ -1,3 +1,4 @@
 # Part of OpenSPP. See LICENSE file for full copyright and licensing details.
 from . import test_hide_menu
 from . import test_migration_dedup_hide_menu
+from . import test_register_hook_guard